% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiR_ovalue.R
\name{HBM_U_upper}
\alias{HBM_U_upper}
\title{Upper confidence bound of mean of U statistics of order 2
via Hoeffding-Bentkus-Maurer inequalities}
\usage{
HBM_U_upper(U, n, alpha)
}
\arguments{
\item{U}{the value of the U statistic}

\item{n}{sample size}

\item{alpha}{confidence level}
}
\description{
Upper confidence bound of mean of U statistics of order 2
via Hoeffding-Bentkus-Maurer inequalities
}
